// frontend/app/lib/api/generated/models/serverSummary.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Internal Server Observation
 * OpenAPI spec version: 0.1.0
 */

export interface ServerSummary {
  active_users?: string[] | null;
  cpu_used_percent?: number | null;
  disk_max_used_percent?: number | null;
  error?: string | null;
  gpu_max_utilization_percent?: number | null;
  host: string;
  is_online: boolean;
  last_updated_at: string;
  memory_used_percent?: number | null;
  name: string;
}